+++ mul.h
@@ -7,11 +7,11 @@ struct MUL{
 };
 struct MUL mul[16];
 void mul_act(struct MUL *mul, int8_t a, int8_t b){
-    mul->c = a * b;
+    mul->c = (int32_t)a * (int32_t)b;
 }
-void mul_store(int32_t (*bram)[16], int row_indx){
+void mul_store(int8_t (*bram)[16], int row_indx){
     for(int i = 0; i < 16; i++){
-        bram[row_indx][i] = mul[i].c;
+        bram[row_indx][i] = (int8_t)mul[i].c;
     }
 }
 #endif
